+++ src/commands/utils/atlas-auth.ts
@@ -1,12 +1,12 @@
-import * as vscode from 'vscode';
-import AxiosDigestAuth from '@mhoc/axios-digest-auth'; 
+// import * as vscode from 'vscode';
+// import AxiosDigestAuth from '@mhoc/axios-digest-auth'; 
 
-const authConfig = vscode.workspace.getConfiguration('demojockey.atlas.auth');
+// const authConfig = vscode.workspace.getConfiguration('demojockey.atlas.auth');
 
-function getPublicKey() { return authConfig.get('public-key', ''); }
-function getPrivateKey() { return authConfig.get('private-key', ''); }
+// function getPublicKey() { return authConfig.get('public-key', ''); }
+// function getPrivateKey() { return authConfig.get('private-key', ''); }
 
-export const axios = new AxiosDigestAuth({
-  username: getPublicKey(),
-  password: getPrivateKey()
-});
+// export const axios = new AxiosDigestAuth({
+//   username: getPublicKey(),
+//   password: getPrivateKey()
+// });